% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augbin_package.R
\docType{package}
\name{augbin-package}
\alias{augbin-package}
\title{augbin: Augmented Binary Method for Analysing Responder Outcomes}
\description{
\strong{augbin} provides functions to aid the use of the augmented binary
method of analysing responder outcomes.
}
\section{Getting started}{


You can install the latest development version of \strong{augbin} from
\href{https://github.com/}{Github} with:

\code{devtools::install_github("mjg211/augbin")}

An introductory example of how to make use of the package's core
functionality can be found \href{https://github.com/mjg211/augbin}{here}.
More detailed support is available in the package vignette, which can be
accessed with \code{vignette("augbin")}. For further help, please contact
James Wason (\email{james.wason@newcastle.ac.uk}) or Michael Grayling
(\email{michael.grayling@newcastle.ac.uk}).
}

\section{Details}{

Currently...

\itemize{
\item \code{\link{augbin}}: ...
}
}

