% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augbin.R
\name{augbin}
\alias{augbin}
\title{Augmented binary method}
\usage{
augbin(data = build_augbin_data(), summary = FALSE)
}
\arguments{
\item{data}{A dataset to analyse. Must be an object of class
\code{"augbin_data"}. See \code{\link[augbin]{build_augbin_data}} for further
details. Defaults to \code{build_augbin_data()}.}

\item{summary}{A \code{\link[base]{logical}} variable indicating whether to
print a summary of the functions progress. Defaults to \code{FALSE}.}
}
\value{
A \code{\link[base]{list}} of additional \code{\link[base]{class}}
\code{"augbin"}, to enable use with several available S3 functions. It will
contain two components:
\itemize{
\item A \code{\link[tibble]{tibble}} in the slot \code{"odds_ratio"} with
three columns, giving the estimated odds ratio and the limits of its 95\%
confidence interval.
\item A \code{\link[base]{list}} in the slot \code{"inputs"}, containing each
of the input variable.
}
}
\description{
Fits the augmented binary method, using an underlying two-component latent
variable model.
}
\details{
The responder outcome is assumed to define a responder when
a continuous component is below a particular dichotomisation threshold and
a binary component is 0.
}
\examples{
# Use the augmented binary method on a randomly generated dataset
odds_ratio <- augbin()
}
\seealso{
\code{\link[augbin]{build_augbin_data}},
\code{\link[augbin]{print.augbin}}, \code{\link[augbin]{summary.augbin}},
\code{\link[augbin]{plot.augbin}}.
}
\author{
James MS Wason, Michael J Grayling.
}
