% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_augbin_data.R
\name{build_augbin_data}
\alias{build_augbin_data}
\title{Build a dataset of class \code{"augbin_data"}}
\usage{
build_augbin_data(continuous = stats::rnorm(100),
  binary = stats::rbinom(100, 1, 0.4), treatment = stats::rbinom(100,
  1, 0.5), dichotomisation = 0, summary = FALSE)
}
\arguments{
\item{continuous}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}
containing the observations of a continuous component of a responder outcome.
Each element should correspond to the observation for a different patient.}

\item{binary}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}
containing the observations of a binary failure component of a responder
outcome. Each element of \code{binary} must be either 0 or 1 and its
\code{\link[base]{length}} should be equal to that of \code{continuous}. In
addition, its elements should be ordered such that the observation for a
particular patient is in the same position as it is in \code{continuous}.}

\item{treatment}{A \code{\link[base]{numeric}} \code{\link[base]{vector}}
containing the treatment assignments. Each element of \code{treatment} must
be either 0 (corresponding to the 'control' arm) or 1 (corresponding to the
'experimental' arm). In addition, its elements should be ordered such that
the treatment assignment for a particular patient is in the same position as
it is in \code{continuous}.}

\item{dichotomisation}{A \code{\link[base]{numeric}} giving the
dichotomisation hreshold for the continuous component (see
\code{continuous}). A patient is a responder if their binary component (see
\code{binary}) is 0 and their continuous component is below
\code{dichotomisation}.}
}
\value{
A \code{\link[tibble]{tibble}}, with additional
\code{\link[base]{class}} \code{"augbin_data"}.
}
\description{
Builds a dataset of class \code{"augbin_data"}, principally for use with
\code{\link[augbin]{augbin}}.
}
\details{
The responder outcome is assumed to define a responder when
a continuous component (see \code{continuous}) is below a particular
dichotomisation threshold (see \code{dichotomisation}) and a binary component
(see \code{binary}) is 0.
}
\examples{
# Create a randomly generated dataset from 100 patients
data <- build_augbin_data()
}
\seealso{
\code{\link[augbin]{augbin}}, \code{\link[augbin]{print.augbin}},
\code{\link[augbin]{summary.augbin}}, \code{\link[augbin]{plot.augbin}}.
}
\author{
Michael J Grayling.
}
