% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.augbin_data.R
\name{summary.augbin_data}
\alias{summary.augbin_data}
\alias{print.augbin_data}
\title{Print a summary of an augbin_data object}
\usage{
\method{summary}{augbin_data}(object, ...)

\method{print}{augbin_data}(object, ...)
}
\description{
Print a summary of an augbin_data object
}
